# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      description: Returns projects scoped by user role. Owners see all, crew see assigned, clients see own.
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProject"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/ValidationError"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProject"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          $ref: "#/components/responses/DeleteSuccess"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/members:
    get:
      operationId: listProjectMembers
      tags: [projects]
      summary: List project members
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: List of project members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMember"
    post:
      operationId: addProjectMember
      tags: [projects]
      summary: Add a member to a project
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectMember"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
        "400":
          $ref: "#/components/responses/ValidationError"
  /projects/{projectId}/members/{userId}:
    delete:
      operationId: removeProjectMember
      tags: [projects]
      summary: Remove a member from a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/DeleteSuccess"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{projectId}/tasks:
    get:
      operationId: listProjectTasks
      tags: [tasks]
      summary: List tasks for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTask"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/ValidationError"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTask"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          $ref: "#/components/responses/DeleteSuccess"
        "404":
          $ref: "#/components/responses/NotFound"
  /tasks/{taskId}/items:
    get:
      operationId: listTaskItems
      tags: [tasks]
      summary: List checklist items for a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of task items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskItem"
    post:
      operationId: createTaskItem
      tags: [tasks]
      summary: Create a checklist item
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskItem"
      responses:
        "201":
          description: Task item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskItem"
        "400":
          $ref: "#/components/responses/ValidationError"
  /task-items/{id}:
    patch:
      operationId: updateTaskItem
      tags: [tasks]
      summary: Update a checklist item
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskItem"
      responses:
        "200":
          description: Task item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskItem"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteTaskItem
      tags: [tasks]
      summary: Delete a checklist item
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          $ref: "#/components/responses/DeleteSuccess"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /reviews/{id}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get a review by ID
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Review details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Update a review
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReview"
      responses:
        "200":
          description: Review updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          $ref: "#/components/responses/DeleteSuccess"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    DeleteSuccess:
      description: Resource deleted
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DeleteResult"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ValidationError:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
      required:
        - error
    DeleteResult:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [lead, active, in_progress, review, delivered, archived]
        phase:
          type: string
          enum: [pre_production, production, post_production, review, delivered]
        organizationId:
          type: string
          nullable: true
        clientId:
          type: string
          nullable: true
        progress:
          type: integer
        dueDate:
          type: string
          format: date-time
          nullable: true
        startDate:
          type: string
          format: date-time
          nullable: true
        budget:
          type: integer
          nullable: true
        thumbnail:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - phase
        - progress
        - createdAt
        - updatedAt
    CreateProject:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [lead, active, in_progress, review, delivered, archived]
        phase:
          type: string
          enum: [pre_production, production, post_production, review, delivered]
        organizationId:
          type: string
        clientId:
          type: string
        progress:
          type: integer
        dueDate:
          type: string
          format: date-time
        startDate:
          type: string
          format: date-time
        budget:
          type: integer
        thumbnail:
          type: string
      required:
        - name
    UpdateProject:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [lead, active, in_progress, review, delivered, archived]
        phase:
          type: string
          enum: [pre_production, production, post_production, review, delivered]
        organizationId:
          type: string
        clientId:
          type: string
        progress:
          type: integer
        dueDate:
          type: string
          format: date-time
        startDate:
          type: string
          format: date-time
        budget:
          type: integer
        thumbnail:
          type: string
    ProjectMember:
      type: object
      properties:
        projectId:
          type: string
        userId:
          type: string
        role:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - projectId
        - userId
        - createdAt
    CreateProjectMember:
      type: object
      properties:
        userId:
          type: string
        role:
          type: string
      required:
        - userId
    Task:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done, blocked]
        assigneeId:
          type: string
          nullable: true
        progress:
          type: integer
        dueDate:
          type: string
          format: date-time
          nullable: true
        sortOrder:
          type: integer
        dependsOnTaskId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - status
        - progress
        - sortOrder
        - createdAt
        - updatedAt
    CreateTask:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done, blocked]
        assigneeId:
          type: string
        progress:
          type: integer
        dueDate:
          type: string
          format: date-time
        sortOrder:
          type: integer
        dependsOnTaskId:
          type: string
      required:
        - title
    UpdateTask:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done, blocked]
        assigneeId:
          type: string
        progress:
          type: integer
        dueDate:
          type: string
          format: date-time
        sortOrder:
          type: integer
        dependsOnTaskId:
          type: string
    TaskItem:
      type: object
      properties:
        id:
          type: string
        taskId:
          type: string
        title:
          type: string
        completed:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - taskId
        - title
        - completed
        - sortOrder
        - createdAt
    CreateTaskItem:
      type: object
      properties:
        title:
          type: string
        completed:
          type: boolean
        sortOrder:
          type: integer
      required:
        - title
    UpdateTaskItem:
      type: object
      properties:
        title:
          type: string
        completed:
          type: boolean
        sortOrder:
          type: integer
...
    Review:
      type: object
      properties:
        id:
          type: string
        deliverableId:
          type: string
        reviewerId:
          type: string
        status:
          type: string
          enum: [pending, approved, revision_requested]
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - deliverableId
        - reviewerId
        - status
        - createdAt
        - updatedAt
...
    UpdateReview:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, revision_requested]
        comment:
          type: string
...