# lib/api-zod/src/generated/types/createTaskItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTaskItem {
  title: string;
  completed?: boolean;
  sortOrder?: number;
}